requirements.txt
influxdb
numpy
pandas
requests
pytz
PyYAML

